+++ R/metadata.R
@@ -1,8 +1,8 @@
 
 #' Nomis metadata concepts and types
 #'
-#' @description Retrieve all concept code options of all Nomis datasets, concept
-#' code options for a given dataset, or the all the options for a given
+#' @description Retrieve all concept code options of all Nomis datasets, 
+#' concept code options for a given dataset, or the all the options for a given
 #' concept variable from a particular dataset. Specifying \code{concept} will
 #' return all the options for a given variable in a particular dataset.
 #'
@@ -26,7 +26,10 @@
 #'
 #' @param additional_queries Any other additional queries to pass to the API.
 #' See \url{https://www.nomisweb.co.uk/api/v01/help} for instructions on
-#' query structure. Defaults to \code{NULL}.
+#' query structure. Defaults to \code{NULL}. Deprecated in package 
+#' versions greater than 0.2.0 and will eventually be removed.
+#' 
+#' @param ... Use to pass any other parameters to the API.
 #'
 #' @seealso \code{\link{nomis_data_info}}
 #' @seealso \code{\link{nomis_get_data}}
@@ -36,7 +39,6 @@
 #' @export
 #'
 #' @examples \donttest{
-#'
 #' a <- nomis_get_metadata('NM_1_1')
 #'
 #' tibble::glimpse(a)
@@ -63,15 +65,23 @@
 #' f <- nomis_get_metadata('NM_1_1', 'item', search = "*married*")
 #'
 #' tibble::glimpse(f)
-#'
-#'
 #' }
 
 nomis_get_metadata <- function(id, concept = NULL,
                                type = NULL, search = NULL,
-                               additional_queries = NULL) {
+                               additional_queries = NULL, ...) {
   if (missing(id)) {
     stop("The dataset ID must be specified.", call. = FALSE)
+  }
+  
+  # Warning message for additional queries
+  if (length(additional_queries) > 0) {
+    additional_query <- additional_queries
+    
+    message("The `additional_query` parameter is
+            deprecated, please use ... instead")
+  } else {
+    additional_query <- NULL
   }
 
   if (is.null(concept)) {
@@ -91,13 +101,29 @@ nomis_get_metadata <- function(id, concept = NULL,
         paste0(search, collapse = ",")
       )
     )
+    
+    dots <- rlang::list2(...) ## eval the dots
+    
+    x <- c()
+    
+    for (i in seq_along(dots)) { # retrieve the dots
+      x[i] <- ifelse(length(dots[[i]]) > 0,
+                     paste0(
+                       "&", names(dots[i]), "=",
+                       paste0(dots[[i]], collapse = ",")
+                     ),
+                     ""
+      )
+    }
+    
+    dots_query <- paste0(x, collapse = "")
 
     df <- tibble::as.tibble(rsdmx::readSDMX(
       paste0(
         base_url, id, "/", concept,
         type_query, "/def.sdmx.xml?",
         search_query,
-        additional_queries
+        additional_queries, dots_query
       )
     ))
   }