+++ R/overview.R
@@ -54,7 +54,6 @@
 #' @seealso \code{\link{nomis_get_metadata}}
 #'
 #' @examples \donttest{
-#'
 #' library(dplyr)
 #'
 #' q <- nomis_overview("NM_1650_1")
@@ -64,7 +63,6 @@
 #' s <- nomis_overview("NM_1650_1", select = c("Units", "Keywords"))
 #'
 #' s %>% tidyr::unnest(name) %>% glimpse()
-#'
 #' }